Remove no-longer-used callback support from camera entity.
Earlier the sensors like fps and failures were used to set the camera
state. But this logic was removed from the tasmota component of the home-assistant.
Removing it from the hatasmota module as well.
Remove related now-unused imports from camera.py.

Minor changes:
It also rearranges the functions in discovery.py so that they are
alphabetical.