Add camera support (#336)
* Added camera support to hatasmota.
It uses the mjpeg stream over the http server to access the camara feed.
This stream implementation is based on the mjpeg integration.

* Fix the linter errors for the hatasmota camera

* Remove unused constants in camera integration.

I had defined them to support the flip/resolution/brightness change
operations.
However, it is easier to support these operations via
automation/blueprints, since these are just commands over mqtt.
Dropping them from the code to avoid the linter warnings.

* Add the missing dependencies in requirements.txt

* Correct the return types of the camera image and stream functions.

This commit also rearranges the code as per the linter suggestions.

* Remove no-longer-used callback support from camera entity.

Earlier the sensors like fps and failures were used to set the camera
state. But this logic was removed from the tasmota component of the home-assistant.
Removing it from the hatasmota module as well.
Remove related now-unused imports from camera.py.

Minor changes:
It also rearranges the functions in discovery.py so that they are
alphabetical.