Camera: Correct the discovery message. (#353)
The entity removal logic in components/tasmota/discovery.py uses a
concatenation of platform string.
Since the discovery message uses 'cam' instead of 'camera', it is
unable to find the entity id during the removal.
Thus, if the tasmota device stops adding "cam":1 in the discovery
message, home assistant keeps seeing the camera entity.
This change corrects that.